Sweetmantech/myc 1068 p2 segment page onclick (#424)
* Docs page for https://api.recoupable.com/api/create_report

* P1: web onSegmentSelect - new report params
    actual:
    The current implementation is passing legacy parameters to the endpoint at api.recoupable.com/api/create_report.
    required:
    Update the client call to pass only the new segmentId parameter and remove any other legacy variables from the call.
    Begin by creating documentation for the updated API endpoint /api/create_report.
    Verify the API docs with me to ensure correctness.
    Once confirmed, update the client-side codebase to align with the documented API endpoint

* Dev logging. remove.

* Remove dev logging.

* P0: onFunnelComplete - zero segments
    actual:
    When the funnel process finishes, the UI shows zero segments. However, after refreshing the page, the correct list of segments is displayed.
    required:
    Update the funnel completion logic so that upon finishing, the correct list of segments is immediately displayed without requiring a page refresh.

* remove dev logging blocking build.

* remove unused import.

* rollback file changes.

* For data fetching components, prefer passing data as props from a parent container rather than fetching directly in presentational components to improve reusability, testing, and maintain single source of truth

* SegmentButton component.

* styling updates

* P2: Segment Page - onClick
        actual:
        Selecting a segment on the segment page currently does nothing, leaving users with no interaction when they click on a segment.
        required:
        Update the segment page so that clicking on a segment triggers the same action as on the funnel analysis complete page—opening a new chat that displays an HTML report of the selected segment.
